+++ plugins/dates/src/01-tagger/index.js
@@ -6,13 +6,20 @@ const methods = [
   require('./04-time'),
   require('./05-shifts'),
   require('./06-intervals'),
-  require('./07-fixup'),
+  require('./07-timezone'),
+  require('./08-fixup'),
 ]
 
 // normalizations to run before tagger
 const normalize = function (doc) {
   // turn '20mins' into '20 mins'
-  doc.numbers().normalize() // this is sorta problematic
+  if (typeof doc.numbers === 'function') {
+    doc.numbers().normalize()
+  } else {
+    console.warn(
+      `Warning: compromise-numbers plugin is not loaded.\n   You should load this plugin \n     - https://bit.ly/3t8RfFG`
+    )
+  }
   return doc
 }