+++ plugins/dates/src/02-ranges/index.js
@@ -6,6 +6,10 @@ const ranges = require('./ranges')
 const parseRange = function (doc, context) {
   // parse-out 'every week ..'
   let interval = parseInterval(doc, context) || {}
+  // if it's *only* an interval response
+  if (doc.found === false) {
+    return Object.assign({}, interval, { start: null, end: null })
+  }
   // try each template in order
   for (let i = 0; i < ranges.length; i += 1) {
     let fmt = ranges[i]
@@ -28,12 +32,21 @@ const parseRange = function (doc, context) {
   }
   let unit = parseDate(doc, context)
   if (unit) {
+    let end = unit.clone().end()
     res = {
       start: unit,
-      end: unit.clone().end(),
+      end: end,
     }
   }
   let combined = Object.assign({}, interval, res)
+  // ensure start is not after end
+  // console.log(combined)
+  if (combined.start && combined.end && combined.start.d.epoch > combined.end.d.epoch) {
+    // console.warn('Warning: Start date is after End date')
+    combined.start = combined.start.start()
+    // combined.end = combined.start.clone()
+  }
+
   return combined
 }
 module.exports = parseRange