+++ plugins/dates/src/find.js
@@ -47,12 +47,16 @@ const findDate = function (doc) {
     dates = dates.splitOn('#WeekDay')
     dates = dates.not('^and')
   }
+  // '5 june, 10 june'
+  m = dates.match('[#Value #Month] #Value #Month', 0)
+  if (m.found) {
+    dates = dates.splitAfter(m)
+  }
   // 'june 5th, june 10th'
   m = dates.match('[#Month #Value] #Month', 0)
   if (m.found) {
     dates = dates.splitAfter(m)
   }
-
   // '20 minutes june 5th'
   m = dates.match('[#Cardinal #Duration] #Date', 0) //but allow '20 minutes ago'
   if (m.found && !dates.has('#Cardinal #Duration] (ago|from|before|after|back)')) {
@@ -63,13 +67,24 @@ const findDate = function (doc) {
   if (m.found) {
     dates = dates.not(m)
   }
-  // // 'january, february'
-  // m = dates.match('^[#Month] (and|or)? #Month$', 0)
-  // if (m.found) {
-  //   dates = dates.splitAfter(m)
-  //   dates = dates.not('^(and|or)')
-  // }
+  // 'one saturday'
+  dates = dates.notIf('^one (#WeekDay|#Month)$')
 
+  // next week tomorrow
+  m = dates.match('(this|next) #Duration [(today|tomorrow|yesterday)]', 0)
+  if (m.found) {
+    dates = dates.splitBefore(m)
+  }
+  // tomorrow 15 march
+  m = dates.match('[(today|tomorrow|yesterday)] #Value #Month', 0)
+  if (m.found) {
+    dates = dates.splitAfter(m)
+  }
+  // tomorrow yesterday
+  m = dates.match('[(today|tomorrow|yesterday)] (today|tomorrow|yesterday)', 0)
+  if (m.found) {
+    dates = dates.splitAfter(m)
+  }
   return dates
 }
 module.exports = findDate