+++ plugins/dates/src/generate.js
@@ -1,5 +1,6 @@
 const spacetime = require('spacetime')
 const maxDate = 8640000000000000
+const max_loops = 500
 
 const shouldPick = function (s, byDay) {
   if (byDay && byDay[s.dayName()] !== true) {
@@ -13,6 +14,10 @@ const generateDates = function (result, context) {
   let list = []
   let max_count = context.max_repeat || 12
   let s = spacetime(result.start || context.today, context.timezone)
+  s = s.startOf('day')
+  if (context.dayStart) {
+    s = s.time(context.dayStart)
+  }
   // should we stop at the end date?
   let end = spacetime(result.end, context.timezone)
   let toAdd = Object.keys(result.repeat.interval)
@@ -28,7 +33,10 @@ const generateDates = function (result, context) {
   // start going!
   let loops = 0
   // TODO: learn how to write better software.
-  while (list.length < max_count && s.epoch < end.epoch) {
+  for (let i = 0; i < max_loops; i += 1) {
+    if (list.length >= max_count || s.epoch >= end.epoch) {
+      break
+    }
     if (shouldPick(s, byDay, end)) {
       list.push(s.iso())
     }
@@ -43,6 +51,12 @@ const generateDates = function (result, context) {
     }
   }
   result.repeat.generated = list
+  // if we got an interval, but not a start/end
+  if (!result.start && result.repeat.generated && result.repeat.generated.length > 1) {
+    let arr = result.repeat.generated
+    result.start = arr[0]
+    result.end = arr[arr.length - 1]
+  }
   return result
 }
 module.exports = generateDates