+++ plugins/dates/src/parseDate/units/_year.js
@@ -33,6 +33,7 @@ class Month extends Unit {
     return this
   }
 }
+
 class AnyQuarter extends Unit {
   constructor(input, unit, context) {
     super(input, unit, context)